tests/wpt/tests/css/cssom-view/window-screen-width-immutable.html

<!DOCTYPE html>
<html>
<head>
    <title>CSSOM View - 4.2 - screen.width immutability</title>
    <link rel="author" title="Neils Christoffersen" href="mailto:neils.christoffersen@gmail.com">
    <link rel="help" href="http://www.w3.org/TR/cssom-view/#the-screen-interface">
    <meta name="flags" content="dom">
    <meta name="assert" content="screen.width is immutable">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <div id="myDiv"></div>
    <div id="log"></div>
    <script>
        var originalVal = window.screen.width;

        // try to set window.screen.width
        window.screen.width = 0;

        // verify window.screen.width didn't change
        test(function() {
            assert_equals(window.screen.width, originalVal, "window.screen.width should be immutable")
        }, "immutability test");
    </script>
</body>
</html>
